Refactore Home page components
